+++ backend/app/routes/__init__.py
@@ -1,0 +1,1 @@
+# Routes package initialization
